requirements.txt
gpxpy==1.4.2
lz4==3.1.1
Pillow==8.1.0
